tout casser mdr colision

+++ globalvar.py
@@ -5,4 +5,4 @@ ecartcase = 2
 nbBriqueX = 20
 nbBriqueY = 15
 
-screenSize = ((nbBriqueX*widthCase)+((nbBriqueX+1)*ecartcase),((nbBriqueY*heightCase)+((nbBriqueY+1)*ecartcase))*3)
+screenSize = (int((nbBriqueX*widthCase)+((nbBriqueX+1)*ecartcase)),int(((nbBriqueY*heightCase)+((nbBriqueY+1)*ecartcase))*3))
